# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Bookify - Booking System API
  version: 1.0.0
  description: Complete API for managing appointments, services, and user authentication
  contact:
    name: API Support
    url: https://github.com/bookify
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local Development Server
  - url: https://api.example.com
    description: Production Server
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/validate:
    post:
      summary: Validate JWT Token
      description: Verify if the current JWT token is valid
      operationId: validateToken
      tags:
        - Authentication
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenValidation'
        '401':
          description: Token is invalid or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/services:
    get:
      summary: List All Services
      description: Get all available services (public endpoint)
      operationId: listServices
      tags:
        - Services
      security: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
        - name: provider_id
          in: query
          description: Filter by provider ID
          schema:
            type: string
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
    post:
      summary: Create Service
      description: Create a new service (provider only)
      operationId: createService
      tags:
        - Services
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
            example:
              name: Dental Checkup
              description: Complete dental examination
              price: 150.00
              duration_minutes: 45
      responses:
        '201':
          description: Service created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only providers can create services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
  schemas:
...
    Service:
      type: object
      required:
        - id
        - provider_id
        - name
        - price
        - duration_minutes
        - is_active
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: "svc_550e8400e29b41d4a716446655440000"
        provider_id:
          type: string
          format: uuid
          example: "usr_550e8400e29b41d4a716446655440001"
        provider_name:
          type: string
          example: "Dr. Jane Smith"
          nullable: true
        provider_email:
          type: string
          format: email
          example: "provider@example.com"
          nullable: true
        name:
          type: string
          example: "Dental Checkup"
        description:
          type: string
          example: "Complete dental examination and consultation"
          nullable: true
        price:
          type: number
          format: double
          example: 150.00
        duration_minutes:
          type: integer
          example: 45
        is_active:
          type: boolean
          example: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
          example: null
        created_at:
          type: string
          format: date-time
          example: "2026-04-20T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-04-20T10:00:00Z"
...
    TokenValidation:
      type: object
      required:
        - valid
        - user_id
      properties:
        valid:
          type: boolean
          example: true
        user_id:
          type: string
          example: "usr_550e8400e29b41d4a716446655440000"
        email:
          type: string
          example: "user@example.com"
        role:
          type: string
          enum: [client, provider, admin]
          example: "client"
    CreateServiceRequest:
      type: object
      required:
        - name
        - price
        - duration_minutes
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "Dental Checkup"
        description:
          type: string
          maxLength: 1000
          example: "Complete dental examination and consultation"
          nullable: true
        price:
          type: number
          format: double
          minimum: 0
          example: 150.00
        duration_minutes:
          type: integer
          minimum: 1
          example: 45
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Validation error"
        message:
          type: string
          example: "Email already exists"
        code:
          type: string
          example: "VALIDATION_ERROR"
          nullable: true
...